Facebook: 0.0.0.4 testing release - Finally reached persistent connection -- silent session refresh + periodical reconnect like original web client - although the connection may fall down, restrictive politics will disconnect when any error occurs (mostly server time-out) - Redesigned session control to fit original web client - Raised default update loop poll rate to 24 seconds -- less traffic, optimal processing load side by side with message loop, which has poll rate of 55 seconds when not active - Work-around for last protocol change (not receiving status message with periodical updates) by loading mobile profile page to get the data (maybe usable for other user details in future) - Rewritten update process to request only changes since last check for updates -- minimal traffic, noticefuly reduced processing load - Completely rewritten contacts handling -- this solves a) avatar flickering b) updating status, real name and idle flag too often - Fixed reloading avatar when not needed - Parsing improvements (solves many issues like buffer overruns/overflows) - Fixed leak of User-Agent strings - Better handling of garbage collecting on shutdown - Possibility to see actual cookies via Options dialog - Optional toggling of logging (default off) - Extra icon for "Mind" function - Light performance improvements in getting data from results - More sophisticated logging system

+++ facebook/JSON_parser/JSON_parser.h
@@ -25,7 +25,7 @@
 
 /* Determine the integer type use to parse non-floating point numbers */
 #if __STDC_VERSION__ >= 199901L || HAVE_LONG_LONG == 1
-typedef long long JSON_int_t;
+typedef __int64 JSON_int_t; // jarvis: Windows fix, originally "long long"
 #define JSON_PARSER_INTEGER_SSCANF_TOKEN "%I64d" // jarvis: Windows fix, originally "%lld"
 #define JSON_PARSER_INTEGER_SPRINTF_TOKEN "%I64d" // jarvis: Windows fix, originally "%lld"
 #else
